refactor, use argparse
